allow to change the timer for delete message

+++ src/commands/rolls/base_roll.ts
@@ -20,6 +20,7 @@ import moment from "moment";
 const t = i18next.getFixedT("en");
 
 export function deleteAfter(message: InteractionResponse | Message, time: number): void {
+	if (time === 0) return;
 	setTimeout(() => {
 		message.delete();
 	}, time);
@@ -78,7 +79,7 @@ export const newScene = {
 				.setDescriptionLocalizations(cmdLn("scene.time.description"))
 				.setRequired(false)
 		),
-	async execute(interaction: CommandInteraction): Promise<void> {
+	async execute(interaction: CommandInteraction, client: EClient): Promise<void> {
 		if (!interaction.guild) return;
 		const allCommands = await interaction.guild.commands.fetch();
 		const channel = interaction.channel;
@@ -117,7 +118,8 @@ export const newScene = {
 
 			const threadMention = channelMention(newThread.id);
 			const msgReply = await reply(interaction,{ content: ul("scene.interaction", {scene: threadMention}) });
-			deleteAfter(msgReply, 180000);
+			const time = client.settings.get(interaction.guild.id, "deleteAfter") ?? 180000;
+			deleteAfter(msgReply, time);
 			const rollID = allCommands.findKey(command => command.name === "roll");
 			const msgToEdit = await newThread.send("_ _");
 			const msg = `${userMention(interaction.user.id)} - <t:${moment().unix()}:R>\n${ul("scene.underscore")} ${scene}\n*roll: </roll:${rollID}>*`;
